+++ inst/include/simmer/process/manager.h
@@ -22,41 +22,52 @@
 
 namespace simmer {
 
+  template <typename T>
   class Manager : public Process {
-    typedef Fn<void(int)> Setter;
+    typedef Fn<void(T)> Setter;
 
   public:
-    Manager(Simulator* sim, const std::string& name, const VEC<double>& duration,
-            const VEC<int>& value, int period, const Setter& set)
-      : Process(sim, name, false, PRIORITY_MANAGER),
-        duration(duration), value(value), period(period), set(set), index(0) {}
+    Manager(Simulator* sim, const std::string& name,
+            const VEC<double>& duration, const VEC<T>& value, int period,
+            const Setter& set, const OPT<T>& init = NONE)
+      : Process(sim, name, false, PRIORITY_MANAGER), duration(duration),
+        value(value), period(period), set(set), init(init), index(0) { reset(); }
 
-    void reset() { index = 0; }
+    void reset() {
+      index = 0;
+      if (init && (duration.empty() || duration[index]))
+        set(*init);
+    }
 
     void run() {
       if (sim->verbose)
-        sim->print("manager", name, MakeString() << value[index]);
+        sim->print("manager", name, "value", MakeString() << value[index]);
 
       set(value[index]);
-      index++;
-      if (index == duration.size()) {
+
+      if (++index == duration.size()) {
         if (period < 0)
           goto end;
         index = 1;
       }
 
       sim->schedule(duration[index], this, priority);
-      end:
-        return;
+    end:
+      return;
     }
 
-    bool activate(double delay = 0) { return Process::activate(duration[index]); }
+    bool activate(double delay = 0) {
+      if (duration.empty())
+        return false;
+      return Process::activate(duration[index]);
+    }
 
   private:
     VEC<double> duration;
-    VEC<int> value;
+    VEC<T> value;
     int period;
     Setter set;
+    OPT<T> init;
     size_t index;
   };