+++ src/RcppExports.cpp
@@ -1249,18 +1249,35 @@ BEGIN_RCPP
 END_RCPP
 }
 // add_resource_manager_
-bool add_resource_manager_(SEXP sim_, const std::string& name, const std::string& param, const std::vector<double>& intervals, const std::vector<int>& values, int period);
-RcppExport SEXP _simmer_add_resource_manager_(SEXP sim_SEXP, SEXP nameSEXP, SEXP paramSEXP, SEXP intervalsSEXP, SEXP valuesSEXP, SEXP periodSEXP) {
+bool add_resource_manager_(SEXP sim_, const std::string& name, const std::string& param, int init, const std::vector<double>& intervals, const std::vector<int>& values, int period);
+RcppExport SEXP _simmer_add_resource_manager_(SEXP sim_SEXP, SEXP nameSEXP, SEXP paramSEXP, SEXP initSEXP, SEXP intervalsSEXP, SEXP valuesSEXP, SEXP periodSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< SEXP >::type sim_(sim_SEXP);
     Rcpp::traits::input_parameter< const std::string& >::type name(nameSEXP);
     Rcpp::traits::input_parameter< const std::string& >::type param(paramSEXP);
+    Rcpp::traits::input_parameter< int >::type init(initSEXP);
     Rcpp::traits::input_parameter< const std::vector<double>& >::type intervals(intervalsSEXP);
     Rcpp::traits::input_parameter< const std::vector<int>& >::type values(valuesSEXP);
     Rcpp::traits::input_parameter< int >::type period(periodSEXP);
-    rcpp_result_gen = Rcpp::wrap(add_resource_manager_(sim_, name, param, intervals, values, period));
+    rcpp_result_gen = Rcpp::wrap(add_resource_manager_(sim_, name, param, init, intervals, values, period));
+    return rcpp_result_gen;
+END_RCPP
+}
+// add_global_manager_
+bool add_global_manager_(SEXP sim_, const std::string& key, double init, const std::vector<double>& intervals, const std::vector<double>& values, int period);
+RcppExport SEXP _simmer_add_global_manager_(SEXP sim_SEXP, SEXP keySEXP, SEXP initSEXP, SEXP intervalsSEXP, SEXP valuesSEXP, SEXP periodSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type sim_(sim_SEXP);
+    Rcpp::traits::input_parameter< const std::string& >::type key(keySEXP);
+    Rcpp::traits::input_parameter< double >::type init(initSEXP);
+    Rcpp::traits::input_parameter< const std::vector<double>& >::type intervals(intervalsSEXP);
+    Rcpp::traits::input_parameter< const std::vector<double>& >::type values(valuesSEXP);
+    Rcpp::traits::input_parameter< int >::type period(periodSEXP);
+    rcpp_result_gen = Rcpp::wrap(add_global_manager_(sim_, key, init, intervals, values, period));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -1377,7 +1394,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_simmer_add_generator_", (DL_FUNC) &_simmer_add_generator_, 8},
     {"_simmer_add_dataframe_", (DL_FUNC) &_simmer_add_dataframe_, 11},
     {"_simmer_add_resource_", (DL_FUNC) &_simmer_add_resource_, 8},
-    {"_simmer_add_resource_manager_", (DL_FUNC) &_simmer_add_resource_manager_, 6},
+    {"_simmer_add_resource_manager_", (DL_FUNC) &_simmer_add_resource_manager_, 7},
+    {"_simmer_add_global_manager_", (DL_FUNC) &_simmer_add_global_manager_, 6},
     {"_simmer_record_ongoing_", (DL_FUNC) &_simmer_record_ongoing_, 2},
     {NULL, NULL, 0}
 };